useThrottledFunction: Throw if `delay` parameter changes
## Problem

`useThrottledFunction's` current implementation doesn't properly handle the case that the `delay` parameter changes between invocations

`useThrottledFunction` uses the first value of `delay` it receives. If later invocations of `useThrottledFunction` have different values for `delay`, those values will silently be ignored. It's unlikely that this is the behavior the caller is expecting.

## Mitigation

This PR updates `useThrottledFunction` to throw if `delay` changes between invocations. This will make it easy for engineers to realize that `useThrottledFunction` is ignoring their new `delay` value.

## Future

If the use case arises where the `delay` argument needs to change between invocations, we can invest the time to implement proper support for it. In that case, we'd likely want to call `throttledFunction` to create a new throttled function with the new `delay` value.

## Testing

I do not have the dev environment set up so I wasn't able to test this change.