Another attempt at the desired joystick behavior
When the joystick was held at the maximum extent, new "move" events would not
fire, so the imaging control server was not receiving a continuous stream of
pan/tilt commands.

This version creates a 10 Hz timer that spits out the latest pan and tilt
command until an "end" event fires.